requirements.txt
## The following requirements were added by pip freeze:
alembic==1.0.0
atomicwrites==1.1.5
attrs==18.1.0
blinker==1.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Caching==1.4.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Menu==0.7.0
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
genomelink==0.2.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.4
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.3.0
oauthlib==2.1.0
pluggy==0.7.1
py==1.5.4
pytest==3.7.1
python-dateutil==2.7.3
python-editor==1.0.3
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
selenium==3.13.0
six==1.11.0
SQLAlchemy==1.2.10
StringGenerator==0.3.1
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.2.1
